Compare variant index, disregarding order
In the store the variants are stored as a Map. The conversion to a List
of values gives non-deterministic ordering, which is causing a test
failure. Ultimately we probably do care about ordering in a
deterministic way but for now compare the values without caring about
the order (by converting to a set).